Fix calling abs(unsigned)
Casting the substraction result to int
since call of overloaded 'abs(uint)' is ambiguous.

+++ src/servers/app/Screen.cpp
@@ -265,7 +265,7 @@ Screen::_FindBestMode(const display_mode* modes, uint32 count,
 		// TODO: check if these scores make sense
 		int32 diff = 1000 * abs(mode.timing.v_display - height)
 			+ int32(fabs(get_mode_frequency(mode) - frequency) * 10)
-			+ 100 * abs(mode.space - colorSpace);
+			+ 100 * abs((int)(mode.space - colorSpace));
 
 		if (bestIndex == -1 || diff < bestDiff) {
 			bestDiff = diff;
